Potential problem isolated, see description
I have noticed that when looping through and performing mutations (around like 280), the loop is not actually used in the checks, it re-inputs the same acid to perform mutation 4 times. I suspect the issue is that the code is only reporting mutations that happen to fall on a certain offset from the pam, resulting in only a few results.